refactor : user_roadmap 관계 테이블 위치 변경 및 테이블명 변경
