+++ src/main/java/com/example/tily/user/User.java
@@ -1,7 +1,7 @@
 package com.example.tily.user;
 
 import com.example.tily.BaseTimeEntity;
-import com.example.tily.roadmap.UserRoadmapRelation;
+import com.example.tily.roadmap.relation.UserRoadmap;
 import com.example.tily.step.UserStepRelation;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -23,7 +23,7 @@ public class User extends BaseTimeEntity {
     private Long id;
 
     @OneToMany(mappedBy =  "user")
-    private List<UserRoadmapRelation> userRoadmapRelations = new ArrayList<>();
+    private List<UserRoadmap> userRoadmaps = new ArrayList<>();
 
     @OneToMany(mappedBy =  "user")
     private List<UserStepRelation> userStepRelations = new ArrayList<>();
